requirements.txt
langchain
langchain_community
langchain-openai
langchain_pinecone
pinecone-client
python-dotenv
GitPython
